feat: separate faucet into crates (#10)
* feat: introduce `miden-faucet` crate

* chore: remove unused deps

* chore: add docs

* chore: lint

* chore: fix readme

* feat: move files

* chore: format

* chore: remove unused deps

* review: update deploy workflows and crates description

* feat: rename binary back to `miden-faucet`

* feat: move metadata from frontend

* feat: small changes

* review: separate errors

* review: move explorer url to the server

* review: rename crate to `miden-client-lib`

* review: move frontend files

* chore: update metadata docs

* review: address comments
