+++ bin/faucet/frontend/index.html
