+++ bin/faucet/frontend/index.js
