+++ bin/faucet/src/pow/mod.rs
@@ -1,15 +1,17 @@
-use std::collections::{BTreeMap, HashMap};
 use std::num::NonZeroUsize;
 use std::sync::{Arc, Mutex};
 use std::time::{SystemTime, UNIX_EPOCH};
 
 use miden_client::account::AccountId;
-use tokio::time::{Duration, interval};
+use tokio::time::Duration;
 
-use super::challenge::Challenge;
-use crate::server::ApiKey;
-use crate::server::get_pow::PowRequest;
-use crate::server::get_tokens::MintRequestError;
+use crate::pow::api_key::ApiKey;
+use crate::pow::challenge::Challenge;
+use crate::pow::challenge_cache::ChallengeCache;
+
+pub mod api_key;
+pub mod challenge;
+mod challenge_cache;
 
 // POW
 // ================================================================================================
@@ -100,15 +102,12 @@ impl PoW {
         challenge: &str,
         nonce: u64,
         current_time: u64,
-    ) -> Result<(), MintRequestError> {
+    ) -> Result<(), PowError> {
         let challenge = Challenge::decode(challenge, self.secret)?;
 
         // Check timestamp validity
         if challenge.is_expired(current_time, self.config.challenge_lifetime) {
-            return Err(MintRequestError::ExpiredServerTimestamp(
-                challenge.timestamp,
-                current_time,
-            ));
+            return Err(PowError::ExpiredServerTimestamp(challenge.timestamp, current_time));
         }
 
         // Validate the challenge
@@ -116,7 +115,7 @@ impl PoW {
         let valid_api_key = *api_key == challenge.api_key;
         let valid_nonce = challenge.validate_pow(nonce);
         if !(valid_nonce && valid_account_id && valid_api_key) {
-            return Err(MintRequestError::InvalidPoW);
+            return Err(PowError::InvalidPoW);
         }
 
         let mut challenge_cache = self
@@ -126,146 +125,41 @@ impl PoW {
 
         // Check if account has recently submitted a challenge.
         if challenge_cache.has_challenge_for_account(account_id) {
-            return Err(MintRequestError::RateLimited);
+            return Err(PowError::RateLimited);
         }
 
         // Check if the cache already contains the challenge. If not, it is inserted.
         if !challenge_cache.insert_challenge(&challenge) {
-            return Err(MintRequestError::ChallengeAlreadyUsed);
+            return Err(PowError::ChallengeAlreadyUsed);
         }
 
         Ok(())
     }
 }
 
-// CHALLENGE CACHE
-// ================================================================================================
-
-/// A cache that keeps track of the submitted challenges.
-///
-/// The cache is used to check if a challenge has already been submitted for a given account and API
-/// key. It also keeps track of the number of challenges submitted for each API key.
-///
-/// The cache is cleaned up periodically, removing expired challenges.
-#[derive(Clone, Default)]
-struct ChallengeCache {
-    /// Maps challenge timestamp to a tuple of `AccountId` and `ApiKey`.
-    challenges: BTreeMap<u64, Vec<(AccountId, ApiKey)>>,
-    /// Maps API key to the number of submitted challenges.
-    challenges_per_key: HashMap<ApiKey, usize>,
-    /// Maps account id to the number of submitted challenges.
-    account_ids: BTreeMap<AccountId, usize>,
+/// Validated and parsed request for the `PoW` challenge.
+pub struct PowRequest {
+    pub account_id: AccountId,
+    pub api_key: ApiKey,
 }
 
-impl ChallengeCache {
-    /// Inserts a challenge into the cache, updating the number of challenges submitted for the
-    /// account and the API key.
-    ///
-    /// Returns whether the value was newly inserted. That is:
-    /// * If the cache did not previously contain this challenge, `true` is returned.
-    /// * If the cache already contained this challenge, `false` is returned, and the cache is not
-    ///   modified.
-    pub fn insert_challenge(&mut self, challenge: &Challenge) -> bool {
-        let account_id = challenge.account_id;
-        let api_key = challenge.api_key.clone();
-
-        // check if (timestamp, account_id, api_key) is already in the cache
-        let issuers = self.challenges.entry(challenge.timestamp).or_default();
-        if issuers.iter().any(|(id, key)| id == &account_id && key == &api_key) {
-            return false;
-        }
-
-        issuers.push((account_id, api_key.clone()));
-        self.challenges_per_key
-            .entry(api_key)
-            .and_modify(|c| *c = c.saturating_add(1))
-            .or_insert(1);
-        self.account_ids
-            .entry(account_id)
-            .and_modify(|c| *c = c.saturating_add(1))
-            .or_insert(1);
-        true
-    }
-
-    /// Checks if a challenge has been submitted for the given account
-    pub fn has_challenge_for_account(&self, account_id: AccountId) -> bool {
-        self.account_ids.contains_key(&account_id)
-    }
-
-    /// Returns the number of challenges submitted for the given API key.
-    pub fn num_challenges_for_api_key(&self, key: &ApiKey) -> usize {
-        self.challenges_per_key.get(key).copied().unwrap_or(0)
-    }
-
-    /// Cleanup expired challenges and update the number of challenges submitted per API key and
-    /// account id.
-    ///
-    /// # Arguments
-    /// * `current_time` - The current timestamp in seconds since the UNIX epoch.
-    /// * `challenge_lifetime` - The duration during which a challenge is valid.
-    ///
-    /// # Panics
-    /// Panics if any expired challenge has no corresponding entries on the account or API key maps.
-    fn cleanup_expired_challenges(&mut self, current_time: u64, challenge_lifetime: Duration) {
-        // Challenges older than this are expired.
-        let limit_timestamp = current_time - challenge_lifetime.as_secs();
-
-        let valid_challenges = self.challenges.split_off(&limit_timestamp);
-        let expired_challenges = std::mem::replace(&mut self.challenges, valid_challenges);
-
-        for issuers in expired_challenges.into_values() {
-            for (account_id, api_key) in issuers {
-                let remove_api_key = self
-                    .challenges_per_key
-                    .get_mut(&api_key)
-                    .map(|c| {
-                        *c = c.saturating_sub(1);
-                        *c == 0
-                    })
-                    .expect("challenge should have had a key entry");
-                if remove_api_key {
-                    self.challenges_per_key.remove(&api_key);
-                }
-
-                let remove_account_id = self
-                    .account_ids
-                    .get_mut(&account_id)
-                    .map(|c| {
-                        *c = c.saturating_sub(1);
-                        *c == 0
-                    })
-                    .expect("challenge should have had an account entry");
-                if remove_account_id {
-                    self.account_ids.remove(&account_id);
-                }
-            }
-        }
-    }
-
-    /// Run the cleanup task.
-    ///
-    /// The cleanup task is responsible for removing expired challenges from the cache.
-    /// It runs every minute and removes challenges that are no longer valid because of their
-    /// timestamp.
-    pub async fn run_cleanup(
-        cache: Arc<Mutex<Self>>,
-        challenge_lifetime: Duration,
-        cleanup_interval: Duration,
-    ) {
-        let mut interval = interval(cleanup_interval);
-
-        loop {
-            interval.tick().await;
-            let current_time = SystemTime::now()
-                .duration_since(UNIX_EPOCH)
-                .expect("current timestamp should be greater than unix epoch")
-                .as_secs();
-            cache
-                .lock()
-                .expect("challenge cache lock should not be poisoned")
-                .cleanup_expired_challenges(current_time, challenge_lifetime);
-        }
-    }
+/// `PoW` challenge related errors.
+#[derive(Debug, thiserror::Error)]
+pub enum PowError {
+    #[error("server timestamp expired, received: {0}, current time: {1}")]
+    ExpiredServerTimestamp(u64, u64),
+    #[error("invalid POW solution")]
+    InvalidPoW,
+    #[error("account is rate limited")]
+    RateLimited,
+    #[error("challenge already used")]
+    ChallengeAlreadyUsed,
+    #[error("server signatures do not match")]
+    ServerSignaturesDoNotMatch,
+    #[error("invalid challenge")]
+    InvalidChallenge,
+    #[error("API key {0} is invalid")]
+    InvalidApiKey(String),
 }
 
 // TESTS
@@ -372,7 +266,7 @@ mod tests {
         let result =
             pow.submit_challenge(account_id, &api_key, &challenge.encode(), nonce, current_time);
         assert!(result.is_err());
-        assert!(matches!(result.err(), Some(MintRequestError::RateLimited)));
+        assert!(matches!(result.err(), Some(PowError::RateLimited)));
     }
 
     #[tokio::test]
