+++ crates/faucet/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+authors.workspace      = true
+description            = "Miden faucet's logic"
+edition.workspace      = true
+homepage.workspace     = true
+keywords               = ["faucet", "miden", "node"]
+license.workspace      = true
+name                   = "miden-faucet-lib"
+readme                 = "README.md"
+repository.workspace   = true
+rust-version.workspace = true
+version.workspace      = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# Miden dependencies.
+miden-client = { features = ["sqlite", "tonic"], workspace = true }
+
+# External dependencies.
+anyhow    = { workspace = true }
+rand      = { features = ["thread_rng"], workspace = true }
+serde     = { features = ["derive"], version = "1.0" }
+thiserror = { workspace = true }
+tokio     = { features = ["fs"], workspace = true }
+tracing   = { workspace = true }
+url       = { workspace = true }
